- Removed precompiled external dependencies from the repo.
- Replaced gtest with xm for unit testing.
- Removed the overload of MakeCallback<>() that returns a FunctionPtrCallback<>.
- Updated stb_truetype.h to v1.24.
- xr::GfxContext fixes.
- Added xr::BinaryHeap<> and its tests.
- Added xr::ScratchBuffer::UseIndices().
- Changed the precision for milliseconds that xr::State and State::Manager take, to double.
- Changed xr;:Timer::GetUST{,Nano}() to double precision, and updated GameLoop.
- Added xr::XonValue::GetLength().
- Added facilities for structured reading, plus a XON implementation.

+++ xr3/h/xr/Timer.hpp
@@ -26,11 +26,11 @@ public:
 
   ///@return Uninterrupted system time in milliseconds. Unaffected by clock
   /// adjustments.
-  static uint64_t GetUST();
+  static double GetUST();
 
   ///@return Uninterrupted system time in nanoseconds. Unaffected by clock
   /// adjustments.
-  static uint64_t GetUSTNano();
+  static double GetUSTNano();
 };
 
 } // xr
